Fix handling default values when processing 3.x params in Flake8 2.x
comma_separated_callback and normalize_paths_callback incorrectly
assumed that `getattr(parser.values, option.dest, value)` would give the
correct value to transform; however, parser.values is initialized with
the /default/ value of the option.

To fix this, we don't compose callbacks, but instead compose the two
transformations and then either call the user-supplied callback or a
default one that simply stores the value in parser.values.

This dramatically simplified the code (and tests) around callbacks. Two
integrationy tests were added that test the previously incorrect
behavior is fixed for Flake8 2.x.